+++ modules/session/src/main/java/com/nautilus_technologies/tsubakuro/impl/low/sql/ResultSetMock.java
@@ -33,7 +33,7 @@ public class ResultSetMock implements ResultSet {
 	    }
 	    return false;
 	}
-        public long fieldCount() {
+        public int fieldCount() {
 	    return 1;
 	}
 	@Deprecated
